Write project file to final location, output target is empty file.
Write the project file to its final location so that paths to the data are
correct and relative to the final location of the project file. Once the project
file is written, create a `READY_TO_ZIP` output target file (empty) to signal
luigi that it is OK to start zipping things up.