Add style for geomagnetic lat/lon
Should be able to re-use the lat/lon layers style by creating a 'label' colum
via `sql` `ogr2ogr_kwargs`.

This proved to be a confusing task...I was unable to cast the `Contour` field to
a str, and attempts to concatenate `Contour` with other fields or text would
cause an `ogr2ogr` sql error.