Bring docker stack up & down on run & cleanup
No more "* 2 were left pending, among these:" 'errors'.

If Luigi is interuppted during a run, or something...it thinks it still has
workers running tasks on subsequent runs, and the only solution is to bounce the
stack. This is really annoying right now, so lets just bounce the stack on every
run/cleanup!!!!!!