Update access instructions for machguth_etal_massbalance_obs_locations
According to the paper,

> At the time of publication the database contains 2961 readings from 633 unique
measuring points. Figure 1 and Table 2 illustrate the locations and provide
basic information for all 53 measuring sites known to the authors. For seven
sites no data were found or the available data were considered unsuitable. The
readings presently included in the database stem from the remaining 46 sites.