Remove bumpversion
It's unclear how to configure bumpversion to handle release candidates
in addition to dev releases. Tried a few things, got tired of it, then
just manually edited the file. We should switch to a maintained project
for version management if we want those features. `tbump` or
`bump2version` will have clearer documentation and more active
communities.