Use proj4 string instead of EPSG code for sea ice age layers
GDAL automatically replaces "EPSG:3408" with the SRS information associated with
"EPSG:6091" because "EPSG:3408" is deprecated and "EPSG:6091" is indicated as
it's replacement in the epsg.org registry. This association is incorrect becuase
EPSG:6091 is EASE2, which uses the WGS84 datum. EPSG:3408 is the original EASE
grid and uses a different datum.